feat: move binning functions to `utils`
